feat(#489): create separate DataType class
